+++ src/common.ts
@@ -1,0 +1,13 @@
+export enum Direction {
+  N,
+  S,
+  W,
+  E,
+}
+
+export interface Rect {
+  x: number
+  y: number
+  w: number
+  h: number
+}
